requirements.txt
# ML Model
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pyserial>=3.5
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0

# Backend
flask>=3.0.0
flask-sqlalchemy>=3.1.0

# Utilities
requests>=2.31.0
